fixing confilcts
