+++ requirements.txt
@@ -4,7 +4,8 @@ requests~=2.31.0
 python-dotenv~=1.0.0
 firebase-admin
 google-cloud-firestore
-pydantic
+pydantic~=2.1.1
+starlette~=0.27.0
 SQLAlchemy
 pyodbc
 sql
