+++ routers/giscloud.py
@@ -1,46 +1,90 @@
 import os
+import json
 from datetime import datetime
-from typing import Annotated, Type
+# from typing import Annotated,Dict
 from urllib.request import Request
 
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, Request
 
 from dependencies import load_lms_token
 from handlers.lms_requests import DeviceData, LMSRequest
 from handlers.monday_handler import Coordinates, Item, MondayClient
+from handlers.jsc_hanler import AzureDbConnection, ConnectionSettings, Fixture
+
 
 router = APIRouter()
-
-
-@router.get("/sites")
-def get_sites(lms: Annotated[LMSRequest, Depends(load_lms_token)]):
-    result = lms.sites()
-    session_site = lms.session(result[0]["company"])
-    if session_site == "Session Successfully":
-        all_groups = lms.get_all_groups()
-        all_sns = lms.get_all_devices(all_groups[0]["id"])
-        new_sn_data = DeviceData(
-            pole="111111",
-            latitude="31.1532",
-            longitude="34.1234",
-            id_gateway=1,
-        )
-        new_sn = lms.create_device(group_id=1, device_data=new_sn_data.to_json())
-        return {"session_site": session_site, "all_groups": all_groups, "all_sns": all_sns, "new_sn": new_sn}
-    return {"result": result, "message": "Hello World", "session_site": session_site}
-
-
-@router.get("/sites/{site_id}")
-def get_site(site_id: str, lms: Annotated[LMSRequest, Depends(load_lms_token)]):
-    lms.session(site_id)
-
-
-@router.post("/giscloud")
-async def new_item(request: Type[Request]):
+lms_base_url = os.getenv("LMS_API_BASEURL")
+lms_request = LMSRequest(lms_base_url)
+conn_settings = ConnectionSettings(
+    server=os.getenv("DB_HOST"),
+    database=os.getenv("DB_NAME"),
+    username=os.getenv("DB_USER"),
+    password=os.getenv("DB_PASSWORD"),
+)
+db_conn = AzureDbConnection(conn_settings)
+db_conn.connect()
+@router.post("/fixture")
+async def update_lms(request):
     try:
         item_data = await request.json()
         # Extract relevant data from the incoming request payload
+        new_fixture = DeviceData(
+            serial_number=item_data.get("sn_nema"),
+            pole=item_data.get("sn_nema"),
+            latitude=item_data.get("latitude"),
+            longitude=item_data.get("longitude"),
+            id_gateway=14,
+            # idIcon=2
+        )
+        old_sn = item_data.get("old_sn")
+        session_site = lms_request.session("Or Yehuda - Israel")
+        new_sn = lms_request.create_device(group_id=259, device_data=new_fixture.to_json())
+        old_sn_res = lms_request.delete_device(group_id=259, serial_number=old_sn)
+        # result = db_conn.delete_fixture(fixture_name=old_sn)
+        return {"session_site": session_site, "new_sn": new_sn, "old_sn": old_sn_res}
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e)) from e
 
+@router.get("/sites")
+def get_sites():
+    # result = lms_request.sites()
+    session_site = lms_request.session("Or Yehuda - Israel")
+    new_fixture = DeviceData(
+        serial_number="10344104",
+        pole="10344104",
+        latitude="31.15323",
+        longitude="34.12343",
+        id_gateway=14,
+    )
+    new_sn = lms_request.create_device(group_id=259, device_data=new_fixture.to_json())
+    # if session_site == "Session Successfully":
+    all_groups = lms_request.get_all_groups()
+    print(all_groups)
+    return {"message": "Hello World", "session_site": session_site, "new_sn": new_sn}
+
+
+@router.get("/sites/{site_id}")
+def get_site(site_id: str):
+    lms_request.session(site_id)
+    session_site = lms_request.session("Or Yehuda - Israel")
+    new_fixture = DeviceData(
+        serial_number="10344104",
+        pole="10344104",
+        latitude="31.15323",
+        longitude="34.12343",
+        id_gateway=14,
+        # idIcon=2
+    )
+    new_sn = lms_request.create_device(group_id=259, device_data=new_fixture.to_json())
+
+
+@router.post("/giscloud")
+async def new_item(request: Request):
+    try:
+        # request_dict = await  json.loads(request.body)
+        item_data_request = await request.json()
+        item_data = item_data_request.get("data")
+        # Extract relevant data from the incoming request payload
         monday_handler = MondayClient(os.getenv("MONDAY_API_KEY"))
         sn_nema = item_data.get("sn_nema")
         insertion_date = datetime.strptime(item_data.get("date"), "%Y-%m-%d %H:%M:%S")
@@ -76,7 +120,34 @@ async def new_item(request: Type[Request]):
             group_id=group_id,
             item=new_item,
         )
-        return {"message": "Item added to Monday.com", "item_id": item_id}
+
+        try:
+            new_fixture = Fixture(
+                name=sn_nema,
+                latitude=coordinates.lat,
+                longitude=coordinates.long,
+                id_gateway=14,
+            )
+            if db_conn.fixture_exists(sn_nema):
+                fixture_id_res = db_conn.update_fixture(new_fixture, fixture_name=sn_nema)
+            else:
+                fixture_id_res = db_conn.insert_fixture(new_fixture)
+            db_conn.delete_fixture(fixture_name=old_sn)
+            response_data = {
+                "LMS result": "Item added to LMS",
+                "id": fixture_id_res,
+                "Monday message": "Item added to Monday.com",
+                "item_id": item_id,
+                "delete old fixture": "fixture deleted successfully",
+                "fixture_id": old_sn,
+            }
+
+            return response_data
+
+        except Exception as e:
+            print(e)
+        finally:
+            db_conn.disconnect()
 
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e)) from e
